test: build log events from one shared builder
Three suites each carried a byte-identical 20-line node builder over the
parser's counter shapes. Adding one field to `Operation` meant editing the
same mock three times, and a field the parser adds has to be present or the
read throws.

They now share `tests/support/logEvents.ts`. Each suite keeps its own thin
wrapper, which is where they genuinely differ: one mocks the filesystem and
the parse, one builds a bare tree.

The identical `jest.mock("fs")` shim in those suites is left alone — the
factory is hoisted, so sharing it is a different change.
